# Significantly reduce number of branching in Node4 search