Before, the inode_4::find_child compiled to 13 branches. Eliminate almost all of
them, leaving two: the found-not found check and whether the found child is not
past the end of array.

This reduces branch mispredictions 3x on micro_benchmark_node4:

before:

2020-06-02 20:24:08
Running ./micro_benchmark_node4
Run on (8 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.57, 0.57, 0.40
---------------------------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------
full_node4_sequential_insert/100         9.59 us         9.59 us        73007 +4=34 16=0 16^=0 256=0 4=34 48=0 48^=0 4^=0 KPfS=9 L=100 items_per_second=10.4222M/s size=12.6289k
full_node4_sequential_insert/512         56.2 us         56.2 us        12455 +4=171 16=0 16^=0 256=0 4=171 48=0 48^=0 4^=0 KPfS=43 L=512 items_per_second=9.11104M/s size=64.5156k
full_node4_sequential_insert/4096         488 us          488 us         1434 +4=1.365k 16=0 16^=0 256=0 4=1.365k 48=0 48^=0 4^=0 KPfS=341 L=4.096k items_per_second=8.39343M/s size=515.984k
full_node4_sequential_insert/32768       4871 us         4871 us          143 +4=10.923k 16=0 16^=0 256=0 4=10.923k 48=0 48^=0 4^=0 KPfS=2.731k L=32.768k items_per_second=6.72697M/s size=4.03127M
full_node4_sequential_insert/65535      10785 us        10784 us           65 +4=21.845k 16=0 16^=0 256=0 4=21.845k 48=0 48^=0 4^=0 KPfS=5.461k L=65.535k items_per_second=6.07709M/s size=8.06238M
full_node4_random_insert/100             13.5 us         13.5 us        51666 items_per_second=7.38373M/s
full_node4_random_insert/512             70.7 us         70.7 us         9891 items_per_second=7.23937M/s
full_node4_random_insert/4096             634 us          634 us         1106 items_per_second=6.46099M/s
full_node4_random_insert/32768           6411 us         6411 us          111 items_per_second=5.11142M/s
full_node4_random_insert/65535          13860 us        13860 us           50 items_per_second=4.72846M/s
node4_full_scan/100                      3.37 us         3.37 us       206965 items_per_second=29.6855M/s
node4_full_scan/512                      28.1 us         28.1 us        24778 items_per_second=18.1974M/s
node4_full_scan/4096                      254 us          254 us         2759 items_per_second=16.1306M/s
node4_full_scan/32768                    2824 us         2824 us          248 items_per_second=11.6016M/s
node4_full_scan/65535                    5845 us         5845 us          120 items_per_second=11.2127M/s
node4_random_gets/100                    59.1 us         59.2 us        11832 items_per_second=16.9041k/s
node4_random_gets/512                     310 us          310 us         2255 items_per_second=3.22136k/s
node4_random_gets/4096                   2606 us         2608 us          268 items_per_second=383.364/s
node4_random_gets/32768                 22054 us        22082 us           32 items_per_second=45.2854/s
node4_random_gets/65535                 44147 us        44205 us           16 items_per_second=22.6221/s
full_node4_sequential_delete/100         6.94 us         6.93 us       100832 items_per_second=14.426M/s
full_node4_sequential_delete/512         36.1 us         36.1 us        19378 items_per_second=14.1891M/s
full_node4_sequential_delete/4096         322 us          322 us         2177 items_per_second=12.7374M/s
full_node4_sequential_delete/32768       2931 us         2931 us          239 items_per_second=11.18M/s
full_node4_sequential_delete/65535       6200 us         6200 us          113 items_per_second=10.5702M/s
full_node4_random_deletes/100            10.2 us         10.1 us        68863 items_per_second=9.85241M/s
full_node4_random_deletes/512            59.3 us         59.3 us        11813 items_per_second=8.63216M/s
full_node4_random_deletes/4096            594 us          594 us         1178 items_per_second=6.89446M/s
full_node4_random_deletes/32768          6776 us         6776 us          103 items_per_second=4.83563M/s
full_node4_random_deletes/65535         17705 us        17705 us           40 items_per_second=3.70152M/s

 Performance counter stats for './micro_benchmark_node4':

         45,887.89 msec task-clock                #    1.000 CPUs utilized
                43      context-switches          #    0.001 K/sec
                 0      cpu-migrations            #    0.000 K/sec
           342,124      page-faults               #    0.007 M/sec
   156,223,667,946      cycles                    #    3.404 GHz                      (33.33%)
    48,770,427,956      stalled-cycles-frontend   #   31.22% frontend cycles idle     (44.44%)
    34,263,227,125      stalled-cycles-backend    #   21.93% backend cycles idle      (55.55%)
   253,253,953,624      instructions              #    1.62  insn per cycle
                                                  #    0.19  stalled cycles per insn  (66.67%)
    59,135,186,960      branches                  # 1288.688 M/sec                    (66.68%)
       942,923,307      branch-misses             #    1.59% of all branches          (66.68%)
    67,277,691,522      L1-dcache-loads           # 1466.132 M/sec                    (55.51%)
     1,162,078,123      L1-dcache-load-misses     #    1.73% of all L1-dcache hits    (22.21%)
       301,062,078      LLC-loads                 #    6.561 M/sec                    (22.21%)
   <not supported>      LLC-load-misses

      45.905796368 seconds time elapsed

      41.151830000 seconds user
       4.736440000 seconds sys

after:

Load Average: 1.24, 0.45, 0.15
---------------------------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------
full_node4_sequential_insert/100         9.27 us         9.28 us        75289 +4=34 16=0 16^=0 256=0 4=34 48=0 48^=0 4^=0 KPfS=9 L=100 items_per_second=10.7716M/s size=12.6289k
full_node4_sequential_insert/512         48.3 us         48.3 us        14478 +4=171 16=0 16^=0 256=0 4=171 48=0 48^=0 4^=0 KPfS=43 L=512 items_per_second=10.5939M/s size=64.5156k
full_node4_sequential_insert/4096         431 us          431 us         1627 +4=1.365k 16=0 16^=0 256=0 4=1.365k 48=0 48^=0 4^=0 KPfS=341 L=4.096k items_per_second=9.49693M/s size=515.984k
full_node4_sequential_insert/32768       4589 us         4589 us          153 +4=10.923k 16=0 16^=0 256=0 4=10.923k 48=0 48^=0 4^=0 KPfS=2.731k L=32.768k items_per_second=7.1407M/s size=4.03127M
full_node4_sequential_insert/65535      10263 us        10262 us           68 +4=21.845k 16=0 16^=0 256=0 4=21.845k 48=0 48^=0 4^=0 KPfS=5.461k L=65.535k items_per_second=6.38603M/s size=8.06238M
full_node4_random_insert/100             12.7 us         12.8 us        54824 items_per_second=7.83672M/s
full_node4_random_insert/512             63.6 us         63.6 us        10978 items_per_second=8.04502M/s
full_node4_random_insert/4096             542 us          542 us         1293 items_per_second=7.56339M/s
full_node4_random_insert/32768           5516 us         5516 us          128 items_per_second=5.94051M/s
full_node4_random_insert/65535          12142 us        12142 us           58 items_per_second=5.3974M/s
node4_full_scan/100                      3.60 us         3.60 us       194179 items_per_second=27.7468M/s
node4_full_scan/512                      22.9 us         22.9 us        30590 items_per_second=22.3723M/s
node4_full_scan/4096                      220 us          220 us         3180 items_per_second=18.6065M/s
node4_full_scan/32768                    2298 us         2298 us          304 items_per_second=14.2615M/s
node4_full_scan/65535                    4823 us         4823 us          145 items_per_second=13.5886M/s
node4_random_gets/100                    57.4 us         57.7 us        12142 items_per_second=17.3459k/s
node4_random_gets/512                     300 us          301 us         2323 items_per_second=3.31937k/s
node4_random_gets/4096                   2485 us         2500 us          280 items_per_second=400.029/s
node4_random_gets/32768                 21026 us        21131 us           33 items_per_second=47.3235/s
node4_random_gets/65535                 42095 us        42307 us           17 items_per_second=23.6366/s
full_node4_sequential_delete/100         7.43 us         7.43 us        93813 items_per_second=13.4566M/s
full_node4_sequential_delete/512         39.7 us         39.7 us        17531 items_per_second=12.8818M/s
full_node4_sequential_delete/4096         377 us          377 us         1858 items_per_second=10.8696M/s
full_node4_sequential_delete/32768       3502 us         3502 us          200 items_per_second=9.35696M/s
full_node4_sequential_delete/65535       7318 us         7319 us           96 items_per_second=8.95467M/s
full_node4_random_deletes/100            9.22 us         9.20 us        76019 items_per_second=10.8649M/s
full_node4_random_deletes/512            50.6 us         50.6 us        13835 items_per_second=10.1206M/s
full_node4_random_deletes/4096            493 us          493 us         1419 items_per_second=8.30374M/s
full_node4_random_deletes/32768          5861 us         5861 us          120 items_per_second=5.5906M/s
full_node4_random_deletes/65535         15594 us        15594 us           45 items_per_second=4.20245M/s

 Performance counter stats for './micro_benchmark_node4':

         47,248.27 msec task-clock                #    1.000 CPUs utilized
                76      context-switches          #    0.002 K/sec
                 0      cpu-migrations            #    0.000 K/sec
           377,906      page-faults               #    0.008 M/sec
   160,864,252,001      cycles                    #    3.405 GHz                      (33.33%)
    61,338,724,500      stalled-cycles-frontend   #   38.13% frontend cycles idle     (44.45%)
    29,926,939,941      stalled-cycles-backend    #   18.60% backend cycles idle      (55.56%)
   282,308,278,386      instructions              #    1.75  insn per cycle
                                                  #    0.22  stalled cycles per insn  (66.68%)
    58,060,783,358      branches                  # 1228.845 M/sec                    (66.68%)
       326,178,471      branch-misses             #    0.56% of all branches          (66.68%)
    70,710,950,037      L1-dcache-loads           # 1496.583 M/sec                    (55.50%)
     1,276,739,784      L1-dcache-load-misses     #    1.81% of all L1-dcache hits    (22.21%)
       328,145,264      LLC-loads                 #    6.945 M/sec                    (22.21%)
   <not supported>      LLC-load-misses

      47.266516419 seconds time elapsed

      42.112205000 seconds user
       5.136513000 seconds sys

Cachegrind with branch predictor simulator:

Before:

==644824==
==644824== I   refs:      4,597,309,419
==644824== I1  misses:           47,726
==644824== LLi misses:           10,805
==644824== I1  miss rate:          0.00%
==644824== LLi miss rate:          0.00%
==644824==
==644824== D   refs:      1,856,270,419  (1,213,432,341 rd   + 642,838,078 wr)
==644824== D1  misses:       19,199,883  (   13,640,035 rd   +   5,559,848 wr)
==644824== LLd misses:        1,282,243  (      336,428 rd   +     945,815 wr)
==644824== D1  miss rate:           1.0% (          1.1%     +         0.9%  )
==644824== LLd miss rate:           0.1% (          0.0%     +         0.1%  )
==644824==
==644824== LL refs:          19,247,609  (   13,687,761 rd   +   5,559,848 wr)
==644824== LL misses:         1,293,048  (      347,233 rd   +     945,815 wr)
==644824== LL miss rate:            0.0% (          0.0%     +         0.1%  )
==644824==
==644824== Branches:        778,104,527  (  729,533,837 cond +  48,570,690 ind)
==644824== Mispredicts:      43,915,577  (   34,731,164 cond +   9,184,413 ind)
==644824== Mispred rate:            5.6% (          4.8%     +        18.9%   )

After:

==644851==
==644851== I   refs:      5,064,263,818
==644851== I1  misses:           47,612
==644851== LLi misses:           10,833
==644851== I1  miss rate:          0.00%
==644851== LLi miss rate:          0.00%
==644851==
==644851== D   refs:      1,942,741,733  (1,254,808,489 rd   + 687,933,244 wr)
==644851== D1  misses:       20,719,381  (   14,907,453 rd   +   5,811,928 wr)
==644851== LLd misses:        1,252,225  (      308,006 rd   +     944,219 wr)
==644851== D1  miss rate:           1.1% (          1.2%     +         0.8%  )
==644851== LLd miss rate:           0.1% (          0.0%     +         0.1%  )
==644851==
==644851== LL refs:          20,766,993  (   14,955,065 rd   +   5,811,928 wr)
==644851== LL misses:         1,263,058  (      318,839 rd   +     944,219 wr)
==644851== LL miss rate:            0.0% (          0.0%     +         0.1%  )
==644851==
==644851== Branches:        738,953,367  (  687,502,294 cond +  51,451,073 ind)
==644851== Mispredicts:      27,797,475  (   17,722,638 cond +  10,074,837 ind)
==644851== Mispred rate:            3.8% (          2.6%     +        19.6%   )